translation-service/python/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
google-generativeai==0.3.2
confluent-kafka==2.2.0
supabase==2.3.1
python-dotenv==1.0.0